@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -28,9 +27,8 @@ ActiveRecord::Schema.define(version: 20161019144131) do
     t.string   "queue"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.index ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
   end
-
-  add_index "delayed_jobs", ["priority", "run_at"], name: "delayed_jobs_priority", using: :btree
 
   create_table "quota", force: :cascade do |t|
     t.string   "name",       limit: 128,                null: false
@@ -38,10 +36,9 @@ ActiveRecord::Schema.define(version: 20161019144131) do
     t.datetime "updated_at",                            null: false
     t.string   "slug",       limit: 128,                null: false
     t.boolean  "approved",               default: true, null: false
+    t.index ["approved"], name: "index_quota_on_approved", using: :btree
+    t.index ["name"], name: "index_quota_on_name", unique: true, using: :btree
+    t.index ["slug"], name: "index_quota_on_slug", unique: true, using: :btree
   end
-
-  add_index "quota", ["approved"], name: "index_quota_on_approved", using: :btree
-  add_index "quota", ["name"], name: "index_quota_on_name", unique: true, using: :btree
-  add_index "quota", ["slug"], name: "index_quota_on_slug", unique: true, using: :btree
 
 end
